+++ lib/plausible/stats/table_decider.ex
@@ -126,7 +126,8 @@ defmodule Plausible.Stats.TableDecider do
   #   See `time_slots` usage in `Plausible.Stats.SQL.Expression` to understand how this is done.
   @smearable_metrics [:visitors, :visits]
   defp smear_session_metrics({:sessions, metrics} = value, query) do
-    if "time:minute" in query.dimensions or "time:hour" in query.dimensions do
+    if ("time:minute" in query.dimensions or "time:hour" in query.dimensions) and
+         not filtering_on_dimension?(query, "event:goal") do
       # Split metrics into two groups: one with visitors and visits, and the remaining ones
       {smearable_metrics, session_metrics} = Enum.split_with(metrics, &(&1 in @smearable_metrics))